module1/backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.1
selectolax>=0.3.17
readability-lxml>=0.8.1
python-dotenv>=1.0.0
pydantic>=2.0.0
